requirements.txt
dash==2.18.1
openai==1.51.0
python-dotenv==1.0.1
gunicorn==22.0.0
